requirements.txt
Flask
flask_pymongo
torchvision